+++ ewebsock/src/lib.rs
@@ -230,18 +230,3 @@ pub fn ws_connect(url: String, options: Options, on_event: EventHandler) -> Resu
 pub fn ws_receive(url: String, options: Options, on_event: EventHandler) -> Result<()> {
     ws_receive_impl(url, options, on_event)
 }
-
-/// transform uri and options into a request builder
-pub fn into_requester(
-    uri: tungstenite::http::Uri,
-    options: Options,
-) -> tungstenite::client::ClientRequestBuilder {
-    let mut client_request = tungstenite::client::ClientRequestBuilder::new(uri);
-    for (key, value) in options.additional_headers {
-        client_request = client_request.with_header(key, value);
-    }
-    for subprotocol in options.subprotocols {
-        client_request = client_request.with_sub_protocol(subprotocol);
-    }
-    client_request
-}
